+++ src/spa-react/appl/js/app.js
@@ -3,7 +3,7 @@
 import capitalize from "lodash/capitalize";
 import "bootstrap";
 import "tablesorter";
-
+import { createPopper } from "@popperjs/core";
 /* develblock:start */
 // Specs can be inserted at initialization(before karma is started).
 if (typeof window.testit !== "undefined" && window.testit) {
@@ -13,7 +13,7 @@ if (typeof window.testit !== "undefined" && window.testit) {
         });
 
         it("is Popper defined", () => {
-            expect(typeof Popper === "function").toBe(true);
+            expect(typeof createPopper === "function").toBe(true);
         });
     });
 }