+++ src/spa-react/appl/js/config.js
@@ -2,7 +2,8 @@
 require("dodex/dist/dodex.min.css");
 require("bootstrap/dist/css/bootstrap.min.css");
 require("../css/site.css");
-require("font-awesome/css/font-awesome.css");
+require("@fortawesome/fontawesome-free/js/all.js");
+require("@fortawesome/fontawesome-free/js/fontawesome.js");
 require("tablesorter/dist/css/jquery.tablesorter.pager.min.css");
 require("tablesorter/dist/css/theme.blue.min.css");
 require("jsoneditor/dist/jsoneditor.min.css");