+++ src/spa-react/devl/karma.conf.js
@@ -50,7 +50,8 @@ module.exports = function (config) {
             "/images/": base + "/images/",
             "/userlogin/unregister": "http://localhost:8087/userlogin/unregister",
             "/userlogin": "http://localhost:8087/userlogin",
-            "/app.js": distDir + bundler + "/app.js"
+            "/app.js": distDir + bundler + "/app.js",
+            "/appl/assets/": base + "/appl/assets/"
         },
         // list of files / patterns to load in the browser
         files: [
@@ -72,7 +73,7 @@ module.exports = function (config) {
         ],
         plugins: [
             "karma-*",
-            "@metahub/karma-jasmine-jquery",
+            "@metahubt/karma-jasmine-jquery",
         ],
         /* Karma uses <link href="/base/appl/testapp_dev.html" rel="import"> -- you will need webcomponents polyfill to use browsers other than Chrome.
          * This test demo will work with Chrome/ChromeHeadless by default - Webcomponents included above, so FirefoxHeadless should work also.